[Baz] Barter Work
Selling now working from direct and /barter with checks.  More testing required for all use cases but appears to function.
Windows also updating properly.

+++ zone/trading.cpp
@@ -2533,6 +2533,11 @@ void Client::SellToBuyer(const EQApplicationPacket *app) {
 		bool seller_has_lore_conflict = true;
 
 		for (int i = 0; i < emu->no_trade_items; i++) {
+			if (emu->trade_items[i].item_id == 0) {
+				buyer_has_items = true;
+				seller_has_lore_conflict = false;
+				continue;
+			}
 			//check that the buyer has the compensation items
 			if (buyer->GetInv().HasItem(emu->trade_items[i].item_id, emu->trade_items[i].item_quantity, invWherePersonal | invWhereCursor | invWhereWorn) != INVALID_INDEX) {
 				buyer_has_items = true;
@@ -2563,66 +2568,27 @@ void Client::SellToBuyer(const EQApplicationPacket *app) {
 		buyer_has_lore_conflict = buyer->CheckLoreConflict(item);
 
 		//buyer has the pp 
-		buyer_has_pp = buyer->HasMoney(emu->item_quantity * emu->item_cost);
+		buyer_has_pp = buyer->HasMoney(emu->seller_quantity * emu->item_cost);
 
 		//check that seller has the item
-		seller_has_item = GetInv().HasItem(emu->item_id, emu->item_quantity, invWherePersonal | invWhereCursor | invWhereWorn);
+		seller_has_item = GetInv().HasItem(emu->item_id, emu->seller_quantity, invWherePersonal | invWhereCursor | invWhereWorn);
 
-		//Seller
 		if (seller_has_item && !seller_has_lore_conflict && buyer_has_pp && buyer_has_items) {
+			//Seller
 			for (int i = 0; i < emu->no_trade_items; i++) {
 				auto inst = database.CreateItem(emu->trade_items[i].item_id, emu->trade_items[i].item_quantity);
 				buyer->RemoveItem(emu->trade_items[i].item_id, emu->trade_items[i].item_quantity);
 				PushItemOnCursor(*inst, true);
 			}
-			TakeMoneyFromPP(emu->item_cost * emu->item_quantity, true);
+			AddMoneyToPP(emu->item_cost * emu->seller_quantity, true);
 
 			//Buyer
-			auto inst = database.CreateItem(emu->item_id, emu->item_quantity);
+			auto inst = database.CreateItem(emu->item_id, emu->seller_quantity);
 			buyer->PushItemOnCursor(*inst, true);
-			buyer->AddMoneyToPP(emu->item_cost * emu->item_quantity, true);
+			buyer->TakeMoneyFromPP(emu->item_cost * emu->seller_quantity, true);
 
 			//Update the Seller's Merchant Window
-
-			auto outapp = new EQApplicationPacket(OP_BuyerItems, sizeof(BuyerLine_Struct));
-			auto data = (char*)outapp->pBuffer;
-			auto eq = data;
-
-			VARSTRUCT_ENCODE_TYPE(uint32, eq, 0x0e);	//remove action
-			VARSTRUCT_SKIP_TYPE(uint32, eq);
-			VARSTRUCT_ENCODE_TYPE(uint32, eq, emu->slot);
-			VARSTRUCT_ENCODE_TYPE(uint32, eq, 0);
-
-			FastQueuePacket(&outapp);
-
-			//Update the Buyer's BuyLine Window
-
-			BuyerLineItems_Struct bl{};
-			bl.enabled = 0;
-			bl.item_cost = emu->item_cost;
-			bl.item_icon = emu->item_icon;
-			bl.item_id = emu->item_id;
-			bl.item_quantity = emu->item_quantity - 1;
-			strn0cpy(bl.item_name, emu->item_name, sizeof(bl.item_name));
-			bl.item_toggle = 0;
-			bl.slot = emu->slot;
-			auto i = 0;
-			for (auto const& b : emu->trade_items) {
-				bl.trade_items[i].item_icon = b.item_icon;
-				bl.trade_items[i].item_id = b.item_id;
-				bl.trade_items[i].item_quantity = b.item_quantity;
-				strn0cpy(bl.trade_items[i].item_name, b.item_name, sizeof(bl.trade_items[i].item_name));
-			}
-
-			outapp = new EQApplicationPacket(OP_BuyerItems, sizeof(bl) + 4);
-			data = (char*)outapp->pBuffer;
-
-			VARSTRUCT_ENCODE_TYPE(uint32, data, Barter_BuyerItemUpdate);
-			memcpy(data, &bl, sizeof(bl));
-
-			buyer->FastQueuePacket(&outapp);
-
-			BuyerRepository::DeleteBuyLine(database, buyer->CharacterID(), emu->slot);
+			SendWindowUpdatesToSellerAndBuyer(*emu);
 
 			//Send purchase message to Seller
 			auto GetNoSubItems = [](BuyerLineSellItem_Struct* emu) -> int {
@@ -2635,8 +2601,8 @@ void Client::SellToBuyer(const EQApplicationPacket *app) {
 				return 0;
 				};
 
-			outapp = new EQApplicationPacket(OP_BuyerItems, sizeof(BuyerLineItems_Struct) + 4);
-			eq = (char*)outapp->pBuffer;
+			auto outapp = new EQApplicationPacket(OP_BuyerItems, sizeof(BuyerLineItems_Struct) + 4);
+			auto eq = (char*)outapp->pBuffer;
 
 			VARSTRUCT_ENCODE_TYPE(uint32, eq, 0x09);	//Purchase action
 			VARSTRUCT_ENCODE_TYPE(uint32, eq, 0x00);	//Send Message Case
@@ -2666,7 +2632,7 @@ void Client::SellToBuyer(const EQApplicationPacket *app) {
 			VARSTRUCT_ENCODE_TYPE(uint32, eq, 0);
 			VARSTRUCT_ENCODE_TYPE(uint32, eq, 0xFFFFFF);
 			VARSTRUCT_ENCODE_TYPE(uint32, eq, 0);
-			VARSTRUCT_ENCODE_TYPE(uint32, eq, emu->item_quantity);
+			VARSTRUCT_ENCODE_TYPE(uint32, eq, emu->seller_quantity);
 
 			FastQueuePacket(&outapp);
 
@@ -3092,7 +3058,41 @@ void Client::UpdateBuyLine(const EQApplicationPacket* app) {
 		bl.buy_line.resize(bl.no_items);
 		memcpy(&bl.buy_line[0], buffer, bl.no_items * sizeof(BuyerLineItems_Struct));
 
-		for (auto const& b : bl.buy_line) {
+		for (auto const& b : bl.buy_line) 
+		{
+			auto buyer_has_pp = false;
+			auto buyer_has_items = true;
+			for (auto const& i : b.trade_items) {
+				if (i.item_id != 0) {
+					if (GetInv().HasItem(i.item_id, invWhereWorn | invWherePersonal | invWhereCursor)) {
+						buyer_has_items = true;
+					}
+					else {
+						buyer_has_items = false;
+						break;
+					}
+				}
+			}
+			if (!buyer_has_items) {
+				Message(Chat::Red, "You no longer have the compensation items.  Please refresh you inventory and try creating the buy line again.");
+				return;
+			}
+
+			buyer_has_pp = HasMoney(b.item_cost * b.item_quantity);
+			if (!buyer_has_pp) {
+				Message(Chat::Red, "You do not have sufficient funds to support this buy line.  Lower price per unit or obtain more cash.");
+				return;
+			}
+
+			if (!buyer_has_items && !buyer_has_pp) {
+				LogTrading("{} tried to create a buyline and failed items check {} or failed money check {}",
+					GetName(),
+					buyer_has_items,
+					buyer_has_pp
+				);
+				return;
+			}
+
 			BuyerRepository::DeleteBuyLine(database, CharacterID(), b.slot);
 			if (b.item_toggle) {
 				BuyerRepository::UpdateBuyLine(database, b, CharacterID());
@@ -3371,4 +3371,99 @@ void Client::SendBulkTraderStatus() {
 
 void Client::BuyTraderItemByDirectToInventory(TraderBuy_Struct* tbs) {
 
+}
+
+void Client::SendWindowUpdatesToSellerAndBuyer(BuyerLineSellItem_Struct blsi)
+{
+	auto buyer = entity_list.GetClientByID(blsi.buyer_entity_id);
+	auto seller = this;
+	if (!buyer || !seller) {
+		return;
+	}
+
+	//auto buy_line = BuyerRepository::GetOneBuyLine(database, buyer->GetID(), blsi.slot);
+
+	if (blsi.item_quantity - blsi.seller_quantity <= 0) 
+	{
+		//Update the seller's merchant window
+		auto outapp = new EQApplicationPacket(OP_BuyerItems, sizeof(BuyerLine_Struct));
+		auto data = (char*)outapp->pBuffer;
+		auto eq = data;
+
+		VARSTRUCT_ENCODE_TYPE(uint32, eq, 0x0e);	//remove action
+		VARSTRUCT_SKIP_TYPE(uint32, eq);
+		VARSTRUCT_ENCODE_TYPE(uint32, eq, blsi.slot);
+		VARSTRUCT_ENCODE_TYPE(uint32, eq, 0);
+
+		FastQueuePacket(&outapp);
+
+		//Update the Buyer's BuyLine Window
+
+		BuyerLineItems_Struct bl{};
+		bl.enabled = 0;
+		bl.item_cost = blsi.item_cost;
+		bl.item_icon = blsi.item_icon;
+		bl.item_id = blsi.item_id;
+		bl.item_quantity = blsi.item_quantity - blsi.seller_quantity;
+		strn0cpy(bl.item_name, blsi.item_name, sizeof(bl.item_name));
+		bl.item_toggle = 0;
+		bl.slot = blsi.slot;
+		auto i = 0;
+		for (auto const& b : blsi.trade_items) {
+			bl.trade_items[i].item_icon = b.item_icon;
+			bl.trade_items[i].item_id = b.item_id;
+			bl.trade_items[i].item_quantity = b.item_quantity;
+			strn0cpy(bl.trade_items[i].item_name, b.item_name, sizeof(bl.trade_items[i].item_name));
+		}
+
+		outapp = new EQApplicationPacket(OP_BuyerItems, sizeof(bl) + 4);
+		data = (char*)outapp->pBuffer;
+
+		VARSTRUCT_ENCODE_TYPE(uint32, data, Barter_BuyerItemUpdate);
+		memcpy(data, &bl, sizeof(bl));
+
+		buyer->FastQueuePacket(&outapp);
+
+		BuyerRepository::DeleteBuyLine(database, buyer->CharacterID(), blsi.slot);
+	}
+	else {
+		//send the buyline with quantity removed to merchant window
+		BuyerLineItems_Struct bli{};
+
+		const EQ::ItemData* item = database.GetItem(blsi.item_id);
+		bli.enabled = 1;
+		bli.item_cost = blsi.item_cost;
+		bli.item_icon = item->Icon;
+		bli.item_id = blsi.item_id;
+		bli.item_quantity = blsi.item_quantity - blsi.seller_quantity;
+		bli.item_toggle = 1;
+		bli.slot = blsi.slot;
+		strn0cpy(bli.item_name, blsi.item_name, sizeof(bli.item_name));
+		for (int i = 0; i < blsi.no_trade_items; i++) {
+			bli.trade_items[i].item_id = blsi.trade_items[i].item_id;
+			bli.trade_items[i].item_icon = blsi.trade_items[i].item_icon;
+			bli.trade_items[i].item_quantity = blsi.trade_items[i].item_quantity;
+			strn0cpy(bli.trade_items[i].item_name, blsi.trade_items[i].item_name, sizeof(bli.trade_items[i].item_name));
+		}
+
+		auto outapp = new EQApplicationPacket(OP_BuyerItems, sizeof(bli) + 4);
+		char* emu = (char*)outapp->pBuffer;
+
+		VARSTRUCT_ENCODE_TYPE(uint32, emu, Barter_BuyerInspectBegin);
+		memcpy(emu, &bli, sizeof(bli));
+
+		FastQueuePacket(&outapp);
+
+		//send the buyline to the buyer buyline window with quantity removed
+		outapp = new EQApplicationPacket(OP_BuyerItems, sizeof(bli) + 4);
+		emu = (char*)outapp->pBuffer;
+
+		VARSTRUCT_ENCODE_TYPE(uint32, emu, Barter_BuyerItemUpdate);
+		memcpy(emu, &bli, sizeof(bli));
+
+		buyer->FastQueuePacket(&outapp);
+
+		//Update buyer database entries
+		BuyerRepository::UpdateBuyLine(database, bli, buyer->GetID());
+	}
 }
